updated migration i to f

+++ db/migrate/20210201003555_create_bids.rb
@@ -2,7 +2,7 @@ class CreateBids < ActiveRecord::Migration[6.0]
   def change
     create_table :bids do |t|
       t.integer :years
-      t.integer :salary_per_year
+      t.float :salary_per_year
       t.belongs_to :player, null: false, foreign_key: true
 
       t.timestamps
