+++ db/migrate/20210202024105_create_players.rb
@@ -5,7 +5,7 @@ class CreatePlayers < ActiveRecord::Migration[6.0]
       t.string :position
       t.string :real_mlb_team
       t.integer :years
-      t.integer :salary_per_year
+      t.float :salary_per_year
       t.belongs_to :team, null: false, foreign_key: true
 
       t.timestamps
